+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- `partial-recall--clean-up-buffer` now calls `quit-window` instead of
+  `delete-window` which could mean deleting the only window.
+
 ## [v0.6.0]
 
 ### Added